add ApiHost option for ConfigGenerator
